Moved things around to satisfy Clippy lints
